Jan 28

+++ src/FactTrailingZeros.java
@@ -3,7 +3,7 @@ public class FactTrailingZeros {
         int count = 0;
         int pwr = 1;
         while ((Math.pow(5, pwr) <= n)) {
-            count += n / Math.pow(5, pwr);
+            count += (int) (n / Math.pow(5, pwr));
             pwr++;
         }
         return count;
